yolov5-5.0/requirements.txt
# YOLOv5 requirements
# PyTorch - 已手动安装
# torch>=1.7.0
# torchvision>=0.8.1

# Core
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0

# Logging and Plotting
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4

# TensorBoard
tensorboard>=2.4.1

# Export
# coremltools>=4.1
# onnx>=1.9.0
# onnx-simplifier>=0.3.6
# nvidia-pyindex
# nvidia-tensorrt

# Extras
# thop  # FLOPS computation
# pycocotools>=2.0  # COCO mAP

